backend/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
midiutil==1.2.1
mido==1.3.0
numpy==1.26.0
pydantic==2.4.0
python-multipart==0.0.6
cors==1.0.1
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
music21==9.1.0
jinja2==3.1.2
pytest==7.4.3
httpx==0.25.1
